Move config store to application scope
