+++ src/features/card/containers/CardListContainer.spec.tsx
@@ -29,7 +29,7 @@ vi.mock("@/features/card/hooks/useCardMutations", () => ({
   }),
 }));
 
-vi.mock("@/features/settings/hooks/useConfig", () => ({ useConfig: () => mocks.config }));
+vi.mock("@/hooks/useConfig", () => ({ useConfig: () => mocks.config }));
 
 vi.mock("@/query/useRemoteCollections", () => ({
   useRemoteCollections: () => {
