+++ src/features/deck/containers/DeckFormContainer.spec.tsx
@@ -11,7 +11,7 @@ const mocks = vi.hoisted(() => ({
   goToList: vi.fn(),
 }));
 
-vi.mock("@/features/settings/hooks/useConfig", () => ({ useConfig: () => mocks.config }));
+vi.mock("@/hooks/useConfig", () => ({ useConfig: () => mocks.config }));
 
 vi.mock("@/query/useRemoteCollections", () => ({
   useRemoteCollections: () => ({
