+++ src/features/study/hooks/useStudyActions.spec.tsx
@@ -11,7 +11,7 @@ const mocks = vi.hoisted(() => ({
   pendingIds: new Set<CardId>(),
 }));
 
-vi.mock("@/features/settings/hooks/useConfig", () => ({
+vi.mock("@/hooks/useConfig", () => ({
   useConfig: () => {
     if (mocks.state == null) throw new Error("Mock state is not initialized");
     return mocks.state.config;
