+++ src/hooks/useConfig.ts
@@ -1,5 +1,5 @@
 import { useStore } from "zustand";
 
-import { configStore } from "@/features/settings/state/configStore";
+import { configStore } from "@/store/configStore";
 
 export const useConfig = (): ConfigState => useStore(configStore, (state) => state.config);
